nvmm: Port to DragonFly #15: anonymous object management
Implement compat code for NetBSD anonymous object management:
uao_create(), uao_reference() and uao_detach().

The created object should be pageable by default, for example, the
object of guest physical memory.  So choose the default pager to create
the anonymous object.

If the object needs to be wired (e.g., the object for communicating
between kernel and userland), the uvm_map_pageable() can be called to
wire the object.