test/nvmm/demo: Update 'toyvirt' to current libnvmm(3)
Various adjustments to the demo code to make it work again with the
current libnvmm(3) API in NetBSD 9.1.

In addition, add one more return check of nvmm_vcpu_configure() and
improve the logging messages a bit.  Update the example output in
README accordingly.

Tested on NetBSD 9.1.  Porting to DragonFly follows.