nvmm: Implement waits for lwkt_send_ipiq_mask()
Unlike lwkt_send_ipiq(), lwkt_send_ipiq_mask() doesn't have a sequence
number to wait for completion, and a wait mechanism like that would be
very expensive.

Here we choose a simple method.  Just have {vmx,svm}_change_cpu()
decrement a global with an atomic op and issue a wakeup() when it hits
0.  And the callers can just tsleep in a loop until its zero

Credit to Matt Dillon for the patch.