pmap: Add some API routines to help NVMM manage guest memory
Add the following three routines for NVMM to use.  NVMM can use these
routines to manipulate the cpumask for the pmap backing guest physical
memory.

* pmap_add_cpu()
* pmap_del_cpu()
* pmap_del_all_cpus()

NOTE: The scheduler might somtimes overload multiple vCPUs on the same
physical cpu, so operating is not quite as simple as calling
add_cpu/del_cpu in the core vmrun routines.

Credit to Matt Dillon