[REVIEW] kernel: Don't siliently ignore EINVAL from mbrinit() and gptinit()
This EINVAL ignoring code was come from FreeBSD since the initial
import.  The main intention of this change is to make the code a bit
more understandable.

I only see one valid case to ignore EINVAL from mbrinit() and that's for
VN(4) to have no MBR and fallback to use the COMPATIBILITY_SLICE
generated by the kernel on default.  I'll restore this behavior in a
later commit.

ok by dillon.