[REVIEW] kernel: Create /dev/slice-by-label/* alias for GPT partition with label
GPT partition can have a label, so use that to create a device alias
under "/dev/slice-by-label/", similar to the aliases created under
"/dev/by-label/" for disklabels that have a packname (custom label).

Don't reuse the same path as the disklabels to avoid interference,
i.e., an alias may be removed by the other.

The GPT labels are encoded with UTF-16, but we simply replace all
non-ASCII characters with a underscore (_), and the resulting label is
further cleaned to replace unwanted characters with a underscore,
similar to the handling of disk serial number.