acpica: Fix unexpected underflow bug in acpi_cputimer.base
The acpi_timer_construct() might underflow timer->base, and this
underflow was unexpected and not handled by acpi_timer_get_timecount*()
functions.  As a result, they might return an extremely huge time count
(e.g., 18446744073706228405) and thus lead to an utterly wrong uptime
(e.g., 42951days ...).

In order to fix the bug and not further complicate the code, introduce
a separate static variable 'acpi_timer_counter' to track the last
counter read from the timer, and use it to determine the delta count,
and simply add the delta to the 'timer->base' to obtain the current time
count.  In this way, we don't need to worry about the underflow of
'timer->base' and achieve a time count monotonically increasing from the
'oldclock' given to the acpi_timer_construct().

Bug: #3299
Reporter: Aedil (IRC)
Discussed-with: dillon